ressl-mcp/requirements.txt
fastmcp
minio

uvicorn
httpx
python-dotenv
pydantic 